split up prelude into files based on type
